+++ Application/Warcraft3PowerLevels/Entities/Neutrals/Summons/DarkMinion_3.cs
@@ -19,7 +19,7 @@ namespace Warcraft3PowerLevels.Entities.Neutral.Summons
         public int Armor { get; set; } = 2;
         public ArmorTypeEnum ArmorType { get; set; } = ArmorTypeEnum.Medium;
 
-        public int Attack { get; set; } = 21;
+        public double Attack { get; set; } = 21;
         public double AttackTime { get; set; } = 1.8;
         public int Range { get; set; } = 100;
