+++ Application/Warcraft3PowerLevels/Entities/Neutrals/Summons/StormPanda.cs
@@ -19,7 +19,7 @@ namespace Warcraft3PowerLevels.Entities.Neutral.Summons
         public int Armor { get; set; } = 4;
         public ArmorTypeEnum ArmorType { get; set; } = ArmorTypeEnum.Light;
 
-        public int Attack { get; set; } = 32;
+        public double Attack { get; set; } = 32;
         public double AttackTime { get; set; } = 1.5;
         public int Range { get; set; } = 600;
