+++ Application/Warcraft3PowerLevels/Entities/Neutrals/Units/OgreMauler.cs
@@ -14,7 +14,7 @@ namespace Warcraft3PowerLevels.Entities.Neutral.Mercenaries
 
         public AttackTypeEnum AttackType { get; set; } = AttackTypeEnum.Normal;
         public WeaponTypeEnum WeaponType { get; set; } = WeaponTypeEnum.Melee;
-        public int Attack { get; set; } = 33;
+        public double Attack { get; set; } = 33;
         public double AttackTime { get; set; } = 1.6;
         public int Range { get; set; } = 100;
