+++ Application/Warcraft3PowerLevels/Entities/NightElves/Summons/AvatarOfVengeance.cs
@@ -14,7 +14,7 @@ namespace Warcraft3PowerLevels.Entities.NightElves.Summons
 
         public AttackTypeEnum AttackType { get; set; } = AttackTypeEnum.Normal;
         public WeaponTypeEnum WeaponType { get; set; } = WeaponTypeEnum.Ranged;
-        public int Attack { get; set; } = 31;
+        public double Attack { get; set; } = 31;
         public double AttackTime { get; set; } = 1.35;
         public int Range { get; set; } = 500;
