+++ Application/Warcraft3PowerLevels/Entities/NightElves/Units/Chimaera_Siege.cs
@@ -14,7 +14,7 @@ namespace Warcraft3PowerLevels.Entities.NightElves.Units
 
         public AttackTypeEnum AttackType { get; set; } = AttackTypeEnum.Siege;
         public WeaponTypeEnum WeaponType { get; set; } = WeaponTypeEnum.Ranged;
-        public int Attack { get; set; } = 50;
+        public double Attack { get; set; } = 50;
         public double AttackTime { get; set; } = 2.5;
         public int Range { get; set; } = 850;
