+++ Application/Warcraft3PowerLevels/Entities/NightElves/Units/DruidOfTheTalon_Crow.cs
@@ -14,7 +14,7 @@ namespace Warcraft3PowerLevels.Entities.NightElves.Units
 
         public AttackTypeEnum AttackType { get; set; } = AttackTypeEnum.Magic;
         public WeaponTypeEnum WeaponType { get; set; } = WeaponTypeEnum.Melee;
-        public int Attack { get; set; } = 14;
+        public double Attack { get; set; } = 14;
         public double AttackTime { get; set; } = 1.6;
         public int Range { get; set; } = 100;
