+++ Application/Warcraft3PowerLevels/Entities/Orcs/Summons/SpiritWolf_3.cs
@@ -14,7 +14,7 @@ namespace Warcraft3PowerLevels.Entities.Orcs.Summons
 
         public AttackTypeEnum AttackType { get; set; } = AttackTypeEnum.Normal;
         public WeaponTypeEnum WeaponType { get; set; } = WeaponTypeEnum.Melee;
-        public int Attack { get; set; } = 22;
+        public double Attack { get; set; } = 22;
         public double AttackTime { get; set; } = 1.0;
         public int Range { get; set; } = 100;
